requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
ffmpeg-python==0.2.0
future==1.0.0
gTTS==2.5.1
idna==3.7
pillow==10.3.0
requests==2.31.0
urllib3==2.2.1
